+++ packages/design-system/src/lib/components/select-advanced/SelectAdvanced.tsx
@@ -18,7 +18,7 @@ import { useIsFirstRender } from './useIsFirstRender'
 export const DEFAULT_LOCALES = { select: 'Select...' }
 export const SELECT_MENU_SEARCH_DEBOUNCE_TIME = 400
 
-export type Option = { value: string; label: string }
+export type Option = { value: string; label: string; group?: string }
 
 export type InputOptions = string[] | Option[]
 
@@ -29,6 +29,7 @@ type BaseProps = {
   isInvalid?: boolean
   inputButtonId?: string
   locales?: { select?: string }
+  hidePlaceholderOption?: boolean
 }
 
 type SingleProps = BaseProps & {
@@ -56,7 +57,8 @@ export const SelectAdvanced = forwardRef(
       isDisabled = false,
       isInvalid = false,
       inputButtonId,
-      locales
+      locales,
+      hidePlaceholderOption = false
     }: SelectAdvancedProps,
     ref: ForwardedRef<HTMLInputElement | null>
   ) => {
@@ -194,6 +196,7 @@ export const SelectAdvanced = forwardRef(
           isSearchable={isSearchable}
           menuId={menuId}
           selectWord={locales?.select ?? DEFAULT_LOCALES.select}
+          hidePlaceholderOption={hidePlaceholderOption}
         />
       </DropdownBS>
     )