+++ tests/component/sections/shared/citation/CitationDownloadButton.spec.tsx
@@ -2,12 +2,17 @@ import { DatasetRepository } from '@/dataset/domain/repositories/DatasetReposito
 import { CitationDownloadButton } from '../../../../../src/sections/shared/citation/citation-download/CitationDownloadButton'
 import { FormattedCitation } from '@/dataset/domain/models/DatasetCitation'
 import { ViewStyledCitationModal } from '@/sections/shared/citation/citation-download/ViewStyledCitationModal'
+import {
+  CSL_STYLES_BASE_URL,
+  CSL_LOCALES_BASE_URL,
+  clearCslCachesForTests
+} from '@/sections/shared/citation/citation-download/csl/cslStyleFetcher'
 import { WithRepositories } from '@tests/component/WithRepositories'
 
 const datasetRepository: DatasetRepository = {} as DatasetRepository
 const mockCitation: FormattedCitation = {
-  content: 'Mock Citation',
-  contentType: 'text/plain'
+  content: JSON.stringify({ id: 'mock-1', type: 'dataset', title: 'Mock Dataset Title' }),
+  contentType: 'application/json'
 }
 
 describe('CitationDownloadButton', () => {
@@ -16,6 +21,12 @@ describe('CitationDownloadButton', () => {
     cy.window().then((win) => {
       cy.stub(win.URL, 'createObjectURL').returns('mock-url')
       cy.stub(win.URL, 'revokeObjectURL')
+    })
+
+    clearCslCachesForTests()
+    cy.intercept('GET', `${CSL_STYLES_BASE_URL}/*.csl`, { fixture: 'citation/test-style.csl' })
+    cy.intercept('GET', `${CSL_LOCALES_BASE_URL}/locales-en-US.xml`, {
+      fixture: 'citation/locales-en-US.xml'
     })
   })
 
@@ -135,6 +146,36 @@ describe('CitationDownloadButton', () => {
     cy.findByText('View Styled Citation').should('exist')
   })
 
+  it('shows a quick-copy icon next to the Cite Dataset dropdown that copies the chicago-author-date citation', () => {
+    datasetRepository.getDatasetCitationInOtherFormats = cy.stub().resolves(mockCitation)
+
+    cy.window().then((win) => {
+      cy.stub(win.navigator.clipboard, 'writeText').resolves()
+    })
+
+    cy.customMount(
+      <WithRepositories datasetRepository={datasetRepository}>
+        <CitationDownloadButton datasetId="test-dataset" version="1.0" />
+      </WithRepositories>
+    )
+
+    cy.findByRole('button', { name: 'Cite Dataset' }).should('exist')
+
+    cy.findByRole('button', { name: /Copy to clipboard icon/ }).then(($copyButton) => {
+      cy.findByRole('button', { name: 'Cite Dataset' }).then(($citeDatasetButton) => {
+        const position = $copyButton[0].compareDocumentPosition($citeDatasetButton[0])
+        expect(position & Node.DOCUMENT_POSITION_FOLLOWING).to.be.greaterThan(0)
+      })
+    })
+
+    cy.findByRole('button', { name: /Copy to clipboard icon/ }).click()
+
+    cy.window().then((win) => {
+      // eslint-disable-next-line @typescript-eslint/unbound-method
+      cy.wrap(win.navigator.clipboard.writeText).should('be.calledWithMatch', 'Mock Dataset Title')
+    })
+  })
+
   it('handles errors when downloading citation', () => {
     datasetRepository.getDatasetCitationInOtherFormats = cy
       .stub()
@@ -149,23 +190,81 @@ describe('CitationDownloadButton', () => {
     cy.findByText('An error occurred while downloading the citation').should('exist')
   })
 
-  it('opens styled citation modal when View Styled Citation is clicked', () => {
-    datasetRepository.getDatasetCitationInOtherFormats = cy.stub().resolves(mockCitation)
+  it('opens styled citation modal and renders the citation formatted in the default style', () => {
     cy.customMount(
-      <WithRepositories datasetRepository={datasetRepository}>
-        <CitationDownloadButton datasetId="test-dataset" version="1.0" />
-      </WithRepositories>
+      <ViewStyledCitationModal show={true} handleClose={() => {}} citation={mockCitation} />
     )
+
+    cy.findByText('Styled Citation').should('exist')
+    cy.findByText('Select a CSL Style').should('exist')
+    cy.findByText('Mock Dataset Title').should('exist')
+    cy.findByRole('button', { name: /Copy to clipboard icon/ }).should('exist')
+    cy.findByRole('dialog').should('exist')
+  })
+
+  it('groups CSL styles into Common Styles and More Styles sections', () => {
+    cy.customMount(
+      <ViewStyledCitationModal show={true} handleClose={() => {}} citation={mockCitation} />
+    )
+
+    cy.get('#cslStyle').click()
+
+    cy.findByText('Select...').should('not.exist')
+    cy.findByText('Common Styles').should('exist')
+    cy.findByText('More Styles').should('exist')
+    cy.findByRole('option', { name: 'chicago-author-date' }).should('exist')
+    cy.findByRole('option', { name: 'ieee' }).should('exist')
+
+    cy.findByPlaceholderText('Search...').type('apa')
+
+    cy.findByText('Common Styles').should('not.exist')
+    cy.findByText('More Styles').should('not.exist')
+    cy.findByRole('option', { name: 'apa' }).should('exist')
+  })
+
+  it('reformats the citation when a different CSL style is selected', () => {
+    cy.intercept('GET', `${CSL_STYLES_BASE_URL}/apa.csl`, { fixture: 'citation/test-style-b.csl' })
 
     cy.customMount(
       <ViewStyledCitationModal show={true} handleClose={() => {}} citation={mockCitation} />
     )
 
-    cy.findByText('Styled Citation').click()
-    cy.findByText('Select a CSL Style').should('exist')
-    cy.findByText(mockCitation.content).should('exist')
-    cy.findByRole('button', { name: /Copy to clipboard icon/ }).should('exist')
-    cy.findByRole('dialog').should('exist')
+    cy.findByText('Mock Dataset Title').should('exist')
+
+    cy.get('#cslStyle').click()
+    cy.findByPlaceholderText('Search...').type('apa')
+    cy.findByRole('option', { name: 'apa' }).click()
+
+    cy.findByText('STYLE-B: Mock Dataset Title').should('exist')
+  })
+
+  it('hides the bibliography entry number for numbered CSL styles', () => {
+    cy.intercept('GET', `${CSL_STYLES_BASE_URL}/apa.csl`, {
+      fixture: 'citation/numbered-style.csl'
+    })
+
+    cy.customMount(
+      <ViewStyledCitationModal show={true} handleClose={() => {}} citation={mockCitation} />
+    )
+
+    cy.get('#cslStyle').click()
+    cy.findByPlaceholderText('Search...').type('apa')
+    cy.findByRole('option', { name: 'apa' }).click()
+
+    cy.findByText('Mock Dataset Title').should('exist')
+    cy.get('.csl-left-margin').should('exist').and('not.be.visible')
+  })
+
+  it('shows an error message when fetching the CSL style fails', () => {
+    cy.intercept('GET', `${CSL_STYLES_BASE_URL}/*.csl`, { statusCode: 500 })
+
+    cy.customMount(
+      <ViewStyledCitationModal show={true} handleClose={() => {}} citation={mockCitation} />
+    )
+
+    cy.contains('An error occurred while formatting the citation in the selected style').should(
+      'exist'
+    )
   })
 
   it('closes styled citation modal when close is triggered', () => {
@@ -183,6 +282,39 @@ describe('CitationDownloadButton', () => {
     cy.findByRole('dialog').should('exist')
     cy.findByRole('button', { name: /close/i }).click()
     cy.findByRole('dialog').should('not.exist')
+  })
+
+  it('remembers the selected CSL style after closing and reopening the modal', () => {
+    cy.intercept('GET', `${CSL_STYLES_BASE_URL}/apa.csl`, { fixture: 'citation/test-style-b.csl' })
+    datasetRepository.getDatasetCitationInOtherFormats = cy.stub().resolves(mockCitation)
+
+    cy.customMount(
+      <WithRepositories datasetRepository={datasetRepository}>
+        <CitationDownloadButton datasetId="test-dataset" version="1.0" />
+      </WithRepositories>
+    )
+
+    cy.findByRole('button', { name: 'Cite Dataset' }).click()
+    cy.findByText('View Styled Citation').click()
+
+    cy.get('#cslStyle').click()
+    cy.findByPlaceholderText('Search...').type('apa')
+    cy.findByRole('option', { name: 'apa' }).click()
+
+    cy.findByText('Citation in apa style').should('exist')
+    cy.findByText('STYLE-B: Mock Dataset Title').should('exist')
+    cy.findByTestId('toggle-inner-content').should('contain.text', 'apa')
+
+    cy.findByRole('button', { name: /close/i }).click()
+    cy.findByRole('dialog').should('not.exist')
+
+    cy.findByRole('button', { name: 'Cite Dataset' }).click()
+    cy.findByText('View Styled Citation').click()
+
+    cy.findByRole('dialog').should('exist')
+    cy.findByText('Citation in apa style').should('exist')
+    cy.findByTestId('toggle-inner-content').should('contain.text', 'apa')
+    cy.findByText('STYLE-B: Mock Dataset Title').should('exist')
   })
 
   it('handles error when fetching styled citation', () => {